refactor(ui): migrate existing ui components to use cva
