+++ src/components/common/navbar/cart-button.tsx
@@ -1,5 +1,6 @@
-import clsx from 'clsx';
 import Image from 'next/image';
+
+import { cn } from '@/lib/utils/cn';
 
 import { NavbarButton } from './navbar-buttons';
 
@@ -10,7 +11,7 @@ function NumberOfItems({ items }: { items: number }) {
   //TO BE IMPLEMENTED: useCart in parent will provide number of items
   return (
     <div
-      className={clsx(
+      className={cn(
         'absolute -top-2 h-[18px] w-[18px] p-1 text-xs flex items-center justify-center rounded-full aspect-square -right-2 bg-primary text-background font-semibold',
         { '-right-4': items > 99 }
       )}
