+++ src/components/common/sidebar/index.tsx
@@ -1,7 +1,7 @@
 'use client';
-import clsx from 'clsx';
 
 import { useSidebar } from '@/lib/providers/sidebar-provider';
+import { cn } from '@/lib/utils/cn';
 
 import ModalOuter from '../modal/modal-outer';
 import NavigationLinks from '../navigation-links';
@@ -15,7 +15,7 @@ export default function Sidebar() {
         id="sidebar"
         aria-hidden={!isOpen} //inert is not supported at all browsers at the moment of creation of this component, so we are adding aria-hidden as well
         inert={!isOpen}
-        className={clsx(
+        className={cn(
           'z-20 fixed flex flex-col gap-6 top-0 right-0 ease-in transition-transform transform origin-right duration-500 w-screen max-w-[359px] h-screen bg-background p-4 pt-8',
           {
             'translate-x-0': isOpen,
