+++ src/stories/button.stories.tsx
@@ -11,6 +11,9 @@ type StoryProps = ComponentProps<typeof Button>;
 const meta: Meta<StoryProps> = {
   component: Button,
   title: 'Components/Button',
+  parameters: {
+    layout: 'centered',
+  },
   tags: ['autodocs'],
   argTypes: {
     variant: {
